FTR proper implementation & SA MA bypass fix
selection code now deselects if requirements not met anymore.
Also fixed silentAimbot bypassing minAngle. Now, a player is no target if not in minagle.
We could also make it not aim @ silentAimbot if minAngle not met, like it was implemented in normal AB before, but this makes more sense IMO